# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs for Shop Scrape App",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Retrieve products",
                "description": "Fetches a list of products with optional pagination.",
                "operationId": "48cf0348a9cb2ad566b7a4db9c71f0c6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "dataPerPage": {
                                        "description": "Number of items per page",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "page": {
                                        "description": "Page number for pagination",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/products/{product}/retailers": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retrieve retailers for a product",
                "description": "Fetches the retailers associated with a specific product.",
                "operationId": "8f478988c4bb945d58825cdb7f4324c1",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RetailerResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/store": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Stores a new product in the database.",
                "operationId": "90841008e6d2f8cb3655547d2377fb01",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of the product",
                                        "type": "string",
                                        "example": "Laptop"
                                    },
                                    "description": {
                                        "description": "Description of the product",
                                        "type": "string",
                                        "example": "High-performance laptop"
                                    },
                                    "manufacturer_part_number": {
                                        "description": "Manufacturer part number",
                                        "type": "string",
                                        "example": "LAP12345"
                                    },
                                    "pack_size_id": {
                                        "description": "ID of the associated pack size",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{product}": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "description": "Updates an existing product by its ID.",
                "operationId": "1ccf5e8c248bb93dc515c3b3c28af2a4",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID of the product to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Updated title of the product",
                                        "type": "string",
                                        "example": "Gaming Laptop"
                                    },
                                    "description": {
                                        "description": "Updated description",
                                        "type": "string",
                                        "example": "High-performance gaming laptop"
                                    },
                                    "manufacturer_part_number": {
                                        "description": "Updated part number",
                                        "type": "string",
                                        "example": "LAP67890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "description": "Deletes a product by its ID.",
                "operationId": "1a0a353d0842ff4f9347bb66ec53e3ea",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Operation success status",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Success message",
                                            "type": "string",
                                            "example": "Product deleted successfully."
                                        },
                                        "data": {
                                            "description": "Additional response data (empty for delete operation)",
                                            "type": "object",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CurrencyResource": {
                "description": "Currency resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the currency",
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "description": "Currency code",
                        "type": "string",
                        "example": "USD"
                    },
                    "name": {
                        "description": "Currency name",
                        "type": "string",
                        "example": "US Dollar"
                    },
                    "symbol": {
                        "description": "Currency symbol",
                        "type": "string",
                        "example": "$"
                    }
                },
                "type": "object"
            },
...
            "PackSizeResource": {
                "description": "Pack Size resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the pack size",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the pack size",
                        "type": "string",
                        "example": "Small"
                    },
                    "weight": {
                        "description": "Weight of the pack size with unit",
                        "type": "string",
                        "example": "500 g"
                    },
                    "amount": {
                        "description": "Amount of items in the pack",
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "description": "Product resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the product",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Title of the product",
                        "type": "string",
                        "example": "Laptop"
                    },
                    "description": {
                        "description": "Description of the product",
                        "type": "string",
                        "example": "High-performance laptop"
                    },
                    "manufacturer_part_number": {
                        "description": "Manufacturer part number",
                        "type": "string",
                        "example": "LAP12345"
                    },
                    "pack_size": {
                        "$ref": "#/components/schemas/PackSizeResource"
                    },
                    "images": {
                        "description": "List of product images",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductImageResource"
                        }
                    }
                },
                "type": "object"
            },
            "ProductImageResource": {
                "description": "Product image resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the product image",
                        "type": "integer",
                        "example": 1
                    },
                    "file_url": {
                        "description": "URL of the image file",
                        "type": "string",
                        "example": "http://example.com/image.jpg"
                    },
                    "file_name": {
                        "description": "Name of the image file",
                        "type": "string",
                        "example": "image.jpg"
                    }
                },
                "type": "object"
            },
...
            "RetailerResource": {
                "description": "Retailer resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the retailer",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Title or name of the retailer",
                        "type": "string",
                        "example": "Best Buy"
                    },
                    "url": {
                        "description": "URL of the retailer",
                        "type": "string",
                        "example": "http://www.bestbuy.com"
                    },
                    "logo": {
                        "description": "Logo URL of the retailer",
                        "type": "string",
                        "example": "http://www.bestbuy.com/logo.png"
                    },
                    "currency": {
                        "$ref": "#/components/schemas/CurrencyResource"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}